requirements.txt
librosa
soundfile
muda
pyyaml
tqdm
matplotlib
seaborn

